+++ tests/testthat/test-labels.R
@@ -131,21 +131,21 @@ test_that("lbl_glue arguments", {
 
 test_that("lbl_midpoints", {
   lbrk <- brk_res(brk_manual(1:3, rep(TRUE, 3)))
-  expect_equivalent(lbl_midpoint()(lbrk), c("1.5", "2.5"))
+  expect_equivalent(lbl_midpoints()(lbrk), c("1.5", "2.5"))
 })
 
 
-test_that("lbl_midpoint arguments", {
+test_that("lbl_midpoints arguments", {
   lbrk <- brk_res(brk_manual(1:3, rep(TRUE, 3)))
-  expect_equivalent(lbl_midpoint(first = "{r}")(lbrk), c("2", "2.5"))
-  expect_equivalent(lbl_midpoint(last = "{l}")(lbrk), c("1.5", "2"))
+  expect_equivalent(lbl_midpoints(first = "{r}")(lbrk), c("2", "2.5"))
+  expect_equivalent(lbl_midpoints(last = "{l}")(lbrk), c("1.5", "2"))
 
   sbrk <- brk_res(brk_manual(c(1, 2, 2, 3), c(TRUE, TRUE, FALSE, TRUE)))
-  expect_equivalent(lbl_midpoint(single = "[{l}]")(sbrk), c("1.5", "[2]", "2.5"))
+  expect_equivalent(lbl_midpoints(single = "[{l}]")(sbrk), c("1.5", "[2]", "2.5"))
 
   qbrk <- brk_res(brk_quantiles(c(0, 0.5, 1)), x = 0:10)
-  expect_equivalent(lbl_midpoint(fmt = percent)(qbrk), c("25%", "75%"))
-  expect_equivalent(lbl_midpoint(raw = TRUE)(qbrk), c("2.5", "7.5"))
+  expect_equivalent(lbl_midpoints(fmt = percent)(qbrk), c("25%", "75%"))
+  expect_equivalent(lbl_midpoints(raw = TRUE)(qbrk), c("2.5", "7.5"))
 })
 
 
@@ -271,29 +271,31 @@ test_that("lbl_discrete arguments", {
 })
 
 
-test_that("lbl_endpoint", {
+test_that("lbl_endpoints", {
   lbrk <- brk_res(brk_default(c(1, 3, 5)), extend = FALSE)
   expect_equivalent(
-    lbl_endpoint()(lbrk),
+    lbl_endpoints()(lbrk),
     c("1", "3")
   )
   expect_equivalent(
-    lbl_endpoint(left = FALSE)(lbrk),
+    lbl_endpoints(left = FALSE)(lbrk),
     c("3", "5")
   )
 })
 
 
-test_that("lbl_endpoint arguments", {
+test_that("lbl_endpoints arguments", {
   lbrk <- brk_res(brk_default(c(1, 3, 5)), extend = FALSE)
   expect_equivalent(
-    lbl_endpoint(fmt = "%.2f")(lbrk),
+    lbl_endpoints(fmt = "%.2f")(lbrk),
     c("1.00", "3.00")
   )
   expect_equivalent(
-    lbl_endpoint(fmt = percent)(lbrk),
+    lbl_endpoints(fmt = percent)(lbrk),
     c("100%", "300%")
   )
+
+  lifecycle::expect_deprecated(lbl_endpoint()(lbrk))
 })
 
 
@@ -312,11 +314,11 @@ test_that("bug: breaks labels don't produce duplicates", {
 })
 
 
-test_that("bug: lbl_endpoint() works with no format and non-standard breaks", {
+test_that("bug: lbl_endpoints() works with no format and non-standard breaks", {
   expect_error(
-    chop_quantiles(0:10, 0.5, labels = lbl_endpoint())
+    chop_quantiles(0:10, 0.5, labels = lbl_endpoints())
     , NA)
   expect_error(
-    chop_mean_sd(0:10, labels = lbl_endpoint())
+    chop_mean_sd(0:10, labels = lbl_endpoints())
     , NA)
 })